requirements.txt
Django==3.1.2
djangorestframework==3.11.0
django-filter==2.4.0
drf-yasg==1.17.1
psycopg2==2.8.4
django-safedelete==0.5.5
